cloud_functions/service_mesh_layer/requirements.txt
# Function dependencies, for example:
# package>=version
functions-framework==2.0.0
requests
python-dotenv
